% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{adder}
\alias{adder}
\title{A function that adds}
\usage{
adder(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{another number}
}
\value{
a number
}
\description{
A function that adds
}
\examples{
adder(2, 4)
}
