% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data_long}
\alias{import_data_long}
\title{Import data and prepare the processing command}
\usage{
import_data_long(
  file,
  silent = F,
  data.origin = c("diann", "other"),
  experiment.type = c("gw", "dr"),
  ...
)
}
\arguments{
\item{file}{file location (leave out to select manually)}

\item{silent}{prevent next command to be printed}

\item{data.origin}{by which software the data was searched}

\item{experiment.type}{whether the experimental design was group-wise ("gw")
or dose-response ("dr")}

\item{...}{further arguments if data.origin is "other" like <sample_names>}
}
\description{
Import data and prepare the processing command
}
