% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdownTest.R
\docType{package}
\name{pkgdownTest}
\alias{pkgdownTest-package}
\alias{pkgdownTest}
\title{Testing features of using pkgdown}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
