% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_data_dr}
\alias{process_data_dr}
\title{Processs the imported DIA-NN precursor data}
\usage{
process_data_dr(
  data_raw,
  peptide.id = "Stripped.Sequence",
  sample.column = "Run",
  quant.column = "Precursor.Normalised",
  filter.rows,
  min_fraction_per_rep = 1,
  min_reps_per_peptide = 3,
  norm.method = "none",
  sample.names,
  sample.concentrations,
  sample.replicates
)
}
\description{
Processs the imported DIA-NN precursor data
}
